Update README.md
 a space where people can learn more about the true process of our food while seeking answers in sustainability, ethics, and the power of consumerism. 
The recipe portion is mainly a way to draw in more users for functionality, but it can also serve as a way to share great recipes and food options with those who think about food in the same way. It’s about creating a space where people can connect, learn, and make more mindful choices about what they eat.